backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.0
httpx==0.25.0
redis==5.0.1
pydantic==2.12.5
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyyaml==6.0.1
prance==25.4.8.0
jsonschema==4.19.2
allure-pytest==2.13.2
pytest==7.4.3
playwright==1.40.0
tenacity==8.2.3
python-dotenv==1.0.0
httpx>=0.25.0
redis>=5.0.0
psutil>=5.9.0